fix: remove the unsafePerformIO and head usage
The cache is now stored inside the `ShakeExtra`. I'm unsure that's the
right location for it.

This should fix hlint. Also, not having a cache stored as top level
entry may makes it not survive session reinit, which seams sane (for
example, during tests).

Still 3 `-boot` related tests are failing. 2 are however flaky.